feat: implement add product to cart feature

+++ app/backend/auth/product.php
@@ -7,7 +7,7 @@ $product_list = $products->pagination(0, 8);
 
 if (Input::exists('POST'))
 {
-  if (Input::get('product_id')) {
+  if (Input::get('product_id') && Input::get('delete_product') == 1 ) {
     $products->delete(Input::get('product_id'));
   }
 }
